move AuditLogService into transmart-server
